requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# LLM and AI
langchain==0.1.4
langchain-groq==0.0.1
groq==0.4.2
openai==1.10.0  # For potential future use

# Document Processing
PyPDF2==3.0.1
python-multipart==0.0.6  # For file uploads

# Utilities
tenacity==8.2.3  # Retry logic
python-dotenv==1.0.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing async endpoints

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0